+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 * Checkboxes in Filter vmenu item
 * Input type range in filter details to eanble / disable filter
 * Add error handling for retrieve single entity details
-* Add spinner to table details and entity childs when waiting for response
+* Add spinner to table details, entity save/delete and entity childs when waiting for response
 
 0.26 (2024-09-11)
 ===============